Expand GetTextWindowPalette(u8 id) to GetTextWindowPalette(u16 id)

+++ include/text_window.h
@@ -24,7 +24,7 @@ void LoadUserWindowBorderGfxOnBg(u8 bg, u16 destOffset, u8 palOffset);
 void DrawTextBorderOuter(u8 windowId, u16 tileNum, u8 palNum);
 void DrawTextBorderInner(u8 windowId, u16 tileNum, u8 palNum);
 void rbox_fill_rectangle(u8 windowId);
-const u16 *GetTextWindowPalette(u8 id);
+const u16 *GetTextWindowPalette(u16 id);
 const u16 *GetOverworldTextboxPalettePtr(void);
 const u32 *GetBattleTextboxPalettePtr(void);
 void LoadSignPostWindowFrameGfx(void);
